Build fix for Windows.
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@1128 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/formats/powertab_old/powertaboldimporter.cpp
@@ -528,7 +528,7 @@ void PowerTabOldImporter::convert(const PowerTabDocument::ChordText &oldChord,
         if (oldName.IsFormulaModificationFlagSet(i))
         {
             name.setModification(static_cast<ChordName::FormulaModification>(
-                std::log(i) / std::log(2)));
+                static_cast<int>(std::log(i) / std::log(2))));
         }
     }
